feat: streaming-prover
There is a sumcheck streaming schedule trait now.
And things work with that.
Next -- integrate the linear time prover -- by computing Az and Bz.
Then do an end to end test -- once this works
Re-organise the rest.
After that see we can have a unified prover -- where the schedule is a
linear time schedule.
So that's all that must change.